Encapsulate `FlatGraph`, separate `FlatGraphBuilder`
