+++ hydroflow_datalog_core/src/lib.rs
@@ -2,7 +2,7 @@ use std::collections::{HashMap, HashSet};
 
 use hydroflow_lang::{
     diagnostic::{Diagnostic, Level},
-    graph::flat_graph::{FlatGraph, FlatGraphBuilder},
+    graph::{FlatGraph, FlatGraphBuilder},
     parse::{IndexInt, Indexing, Pipeline, PipelineLink},
 };
 use proc_macro2::{Span, TokenStream};
